Refactor MCP parity tools onto shared services
