+++ src/awf/api/routes/artifacts.py
@@ -3,40 +3,36 @@
 from __future__ import annotations
 
 import asyncio
-from datetime import UTC, datetime
-from hashlib import sha256
-from pathlib import Path
 
 from fastapi import APIRouter, Depends, HTTPException, status
-from pydantic import BaseModel
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from awf.api.deps import get_db_session, require_api_token
-from awf.common.config import get_settings
+from awf.api.schemas import WorkspaceArtifactListResponse, WorkspaceArtifactResponse
 from awf.db.repositories import WorkspaceRepository
+from awf.service.artifacts import (
+    DEFAULT_ARTIFACT_LIST_LIMIT,
+    _artifact_id,
+    _artifact_kind,
+    _list_artifacts,
+    _workspace_artifact_dir,
+    list_workspace_artifacts_metadata,
+)
 
 router = APIRouter(prefix="/v1/workspaces/{workspace_id}/artifacts", tags=["artifacts"])
 
-DEFAULT_ARTIFACT_LIST_LIMIT = 50
-
-
-class WorkspaceArtifactResponse(BaseModel):
-    artifact_id: str
-    workspace_id: str
-    name: str
-    relative_path: str
-    path: str
-    kind: str
-    size_bytes: int
-    modified_at: datetime
-
-
-class WorkspaceArtifactListResponse(BaseModel):
-    items: list[WorkspaceArtifactResponse]
-    next_cursor: str | None = None
-    has_more: bool = False
-    limit: int = DEFAULT_ARTIFACT_LIST_LIMIT
-    cursor: str | None = None
+__all__ = [
+    "DEFAULT_ARTIFACT_LIST_LIMIT",
+    "WorkspaceArtifactListResponse",
+    "WorkspaceArtifactResponse",
+    "_artifact_id",
+    "_artifact_kind",
+    "_list_artifacts",
+    "_require_workspace",
+    "_workspace_artifact_dir",
+    "asyncio",
+    "list_workspace_artifacts",
+]
 
 
 @router.get(
@@ -48,71 +44,22 @@ async def list_workspace_artifacts(
     workspace_id: str,
     session: AsyncSession = Depends(get_db_session),
 ) -> WorkspaceArtifactListResponse:
-    await _require_workspace(session, workspace_id)
-    artifact_dir = _workspace_artifact_dir(workspace_id)
-    items = await asyncio.to_thread(_list_artifacts, workspace_id, artifact_dir)
-    return WorkspaceArtifactListResponse(
-        items=items,
-        limit=DEFAULT_ARTIFACT_LIST_LIMIT,
-        cursor=None,
+    response = await list_workspace_artifacts_metadata(
+        session,
+        workspace_id=workspace_id,
     )
+    if response is None:
+        raise _workspace_not_found(workspace_id)
+    return response
 
 
 async def _require_workspace(session: AsyncSession, workspace_id: str) -> None:
     if not await WorkspaceRepository(session).exists(workspace_id):
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail={"error_code": "NOT_FOUND", "message": f"No workspace with id {workspace_id}"},
-        )
+        raise _workspace_not_found(workspace_id)
 
 
-def _workspace_artifact_dir(workspace_id: str) -> Path:
-    return Path(get_settings().work_dir) / "artifacts" / workspace_id
-
-
-def _list_artifacts(workspace_id: str, artifact_dir: Path) -> list[WorkspaceArtifactResponse]:
-    try:
-        if not artifact_dir.is_dir() or artifact_dir.is_symlink():
-            return []
-        root = artifact_dir.resolve(strict=True)
-    except OSError:
-        return []
-
-    items: list[WorkspaceArtifactResponse] = []
-    for directory, dirnames, filenames in artifact_dir.walk(follow_symlinks=False):
-        dirnames.sort()
-        for filename in sorted(filenames):
-            candidate = directory / filename
-            try:
-                if candidate.is_symlink():
-                    continue
-                resolved = candidate.resolve(strict=True)
-                if not resolved.is_file() or not resolved.is_relative_to(root):
-                    continue
-                stat = resolved.stat()
-            except OSError:
-                continue
-            relative_path = candidate.relative_to(artifact_dir).as_posix()
-            items.append(
-                WorkspaceArtifactResponse(
-                    artifact_id=_artifact_id(workspace_id, relative_path),
-                    workspace_id=workspace_id,
-                    name=candidate.name,
-                    relative_path=relative_path,
-                    path=str(resolved),
-                    kind=_artifact_kind(candidate),
-                    size_bytes=stat.st_size,
-                    modified_at=datetime.fromtimestamp(stat.st_mtime, UTC),
-                )
-            )
-    return sorted(items, key=lambda item: item.relative_path)
-
-
-def _artifact_id(workspace_id: str, relative_path: str) -> str:
-    digest = sha256(f"{workspace_id}\0{relative_path}".encode()).hexdigest()[:24]
-    return f"art_{digest}"
-
-
-def _artifact_kind(path: Path) -> str:
-    suffix = path.suffix.lower().lstrip(".")
-    return suffix or "file"
+def _workspace_not_found(workspace_id: str) -> HTTPException:
+    return HTTPException(
+        status_code=status.HTTP_404_NOT_FOUND,
+        detail={"error_code": "NOT_FOUND", "message": f"No workspace with id {workspace_id}"},
+    )
