+++ src/awf/api/routes/workspaces.py
@@ -11,11 +11,7 @@ returns 409 ``IDEMPOTENCY_CONFLICT`` per docs/PLAN_MVP.md § Error code taxonomy
 from __future__ import annotations
 
 import asyncio
-import base64
-import binascii
-import json
 from collections.abc import Callable
-from dataclasses import dataclass
 from datetime import datetime
 from typing import Annotated, cast
 
@@ -27,23 +23,20 @@ from awf.api.deps import get_db_session
 from awf.api.schemas import (
     ErrorResponse,
     StaleReasonListResponse,
-    StaleReasonResponse,
     WorkspaceAcceptedResponse,
     WorkspaceCreateRequest,
     WorkspaceCreateV2Request,
     WorkspaceEventListResponse,
     WorkspaceEventResponse,
     WorkspaceOverviewListResponse,
-    WorkspaceOverviewResponse,
     WorkspaceResponse,
     WorkspaceRetryResponse,
     WorkspaceWarningResponse,
 )
 from awf.common.config import Settings, get_settings
-from awf.db.enums import AgentRuntime, OperationStatus, WorkspaceStatus
+from awf.db.enums import AgentRuntime, WorkspaceStatus
 from awf.db.models import Workspace
 from awf.db.repositories import (
-    StaleReasonRepository,
     ValidationRunRepository,
     WorkspaceEventRepository,
     WorkspaceRepository,
@@ -55,8 +48,12 @@ from awf.service.validation_observability import (
     validation_freshness_summary,
 )
 from awf.service.workspace_observability import (
-    workspace_events_by_occurrence,
-    workspace_observability_payload,
+    InvalidWorkspaceOverviewCursorError,
+    _decode_overview_cursor,
+    _encode_overview_cursor,
+    _WorkspaceOverviewCursor,
+    list_workspace_overview_response,
+    list_workspace_stale_reasons_response,
 )
 from awf.service.workspaces import (
     WorkspaceRetryError,
@@ -73,15 +70,20 @@ router = APIRouter(prefix="/v1/workspaces", tags=["workspaces"])
 router_v2 = APIRouter(prefix="/v2/workspaces", tags=["workspaces-v2"])
 DiskCheckProvider = Callable[[Settings], DiskCheck]
 
-
-@dataclass(frozen=True)
-class _WorkspaceOverviewCursor:
-    created_at: datetime
-    workspace_id: str
-
-
-class InvalidWorkspaceOverviewCursorError(ValueError):
-    """Raised when a workspace overview pagination cursor cannot be decoded."""
+__all__ = [
+    "InvalidWorkspaceOverviewCursorError",
+    "_WorkspaceOverviewCursor",
+    "_decode_overview_cursor",
+    "_encode_overview_cursor",
+    "create_workspace",
+    "create_workspace_v2",
+    "get_workspace",
+    "list_workspace_events",
+    "list_workspace_overview",
+    "list_workspace_stale_reasons",
+    "list_workspaces",
+    "retry_workspace",
+]
 
 
 @router.post(
@@ -251,7 +253,15 @@ async def list_workspace_overview(
     session: AsyncSession = Depends(get_db_session),
 ) -> WorkspaceOverviewListResponse:
     try:
-        decoded_cursor = _decode_overview_cursor(cursor)
+        return await list_workspace_overview_response(
+            session,
+            workspace_status=workspace_status,
+            agent=agent,
+            repo_url=repo_url,
+            limit=limit,
+            cursor=cursor,
+            workspace_repository_type=WorkspaceRepository,
+        )
     except InvalidWorkspaceOverviewCursorError as exc:
         raise HTTPException(
             status_code=status.HTTP_400_BAD_REQUEST,
@@ -260,73 +270,6 @@ async def list_workspace_overview(
                 "message": "Invalid workspace overview cursor.",
             },
         ) from exc
-    rows = await WorkspaceRepository(session).list(
-        status=workspace_status,
-        agent=agent,
-        repo_url=repo_url,
-        before_created_at=decoded_cursor.created_at if decoded_cursor is not None else None,
-        before_workspace_id=decoded_cursor.workspace_id if decoded_cursor is not None else None,
-        limit=limit + 1,
-    )
-    page_rows = rows[:limit]
-    has_more = len(rows) > limit
-    items: list[WorkspaceOverviewResponse] = []
-    for ws in page_rows:
-        ordered_events = workspace_events_by_occurrence(ws)
-        observability = workspace_observability_payload(
-            ws,
-            ordered_events=ordered_events,
-        )
-        latest_event = ordered_events[-1] if ordered_events else None
-        active_operation = next(
-            (
-                op
-                for op in sorted(ws.operations, key=lambda item: item.created_at, reverse=True)
-                if op.status in {OperationStatus.pending.value, OperationStatus.running.value}
-            ),
-            None,
-        )
-        items.append(
-            WorkspaceOverviewResponse(
-                workspace_id=ws.id,
-                task_id=ws.task_external_id or ws.id,
-                title=ws.task_title,
-                task_prompt=getattr(ws, "task_prompt", ""),
-                repo_url=ws.repo_url,
-                base_branch=ws.branch_base,
-                branch_name=ws.branch_name,
-                task_class=ws.task_class,
-                owned_paths=list(ws.owned_paths),
-                agent=AgentRuntime(ws.agent),
-                agent_model=observability["agent_model"],
-                agent_effort=observability["agent_effort"],
-                agent_model_source=observability["agent_model_source"],
-                agent_effort_source=observability["agent_effort_source"],
-                lifecycle=observability["lifecycle"],
-                llm_usage=observability["llm_usage"],
-                recovery=observability["recovery"],
-                status=WorkspaceStatus(ws.status),
-                current_phase=ws.status,
-                active_operation=active_operation.type if active_operation is not None else None,
-                last_event=(
-                    WorkspaceEventResponse.model_validate(latest_event)
-                    if latest_event is not None
-                    else None
-                ),
-                pr_url=ws.pr_url,
-                failure_reason=ws.failure_reason,
-                failure_message=ws.failure_message,
-                created_at=ws.created_at,
-                updated_at=ws.updated_at,
-            )
-        )
-    return WorkspaceOverviewListResponse(
-        items=items,
-        next_cursor=_encode_overview_cursor(page_rows[-1]) if has_more and page_rows else None,
-        has_more=has_more,
-        limit=limit,
-        cursor=cursor,
-    )
 
 
 @router.get("/{workspace_id}/events", response_model=WorkspaceEventListResponse)
@@ -364,24 +307,17 @@ async def list_workspace_stale_reasons(
     include_resolved: Annotated[bool, Query()] = False,
     session: AsyncSession = Depends(get_db_session),
 ) -> StaleReasonListResponse:
-    repo = WorkspaceRepository(session)
-    if not await repo.exists(workspace_id):
+    response = await list_workspace_stale_reasons_response(
+        session,
+        workspace_id=workspace_id,
+        include_resolved=include_resolved,
+    )
+    if response is None:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail={"error_code": "NOT_FOUND", "message": f"No workspace with id {workspace_id}"},
         )
-
-    stale_repo = StaleReasonRepository(session)
-    rows = (
-        await stale_repo.list_for_workspace(workspace_id)
-        if include_resolved
-        else await stale_repo.list_active_for_workspace(workspace_id)
-    )
-    return StaleReasonListResponse(
-        items=[StaleReasonResponse.model_validate(row) for row in rows],
-        limit=len(rows),
-        cursor=None,
-    )
+    return response
 
 
 @router.post(
@@ -461,40 +397,6 @@ def _accepted(
         accepted_at=created_at,
         warnings=list(warnings or []),
     )
-
-
-def _encode_overview_cursor(workspace: Workspace) -> str:
-    payload = {
-        "t": workspace.created_at.isoformat(),
-        "id": workspace.id,
-    }
-    encoded = base64.urlsafe_b64encode(json.dumps(payload, separators=(",", ":")).encode("utf-8"))
-    return encoded.decode("ascii").rstrip("=")
-
-
-def _decode_overview_cursor(cursor: str | None) -> _WorkspaceOverviewCursor | None:
-    if cursor is None:
-        return None
-    try:
-        padded_cursor = cursor + ("=" * (-len(cursor) % 4))
-        decoded = base64.urlsafe_b64decode(padded_cursor.encode("ascii"))
-        payload = json.loads(decoded.decode("utf-8"))
-        created_at = datetime.fromisoformat(
-            payload["t"] if "t" in payload else payload["created_at"]
-        )
-        workspace_id = payload["id"] if "id" in payload else payload["workspace_id"]
-    except (
-        binascii.Error,
-        KeyError,
-        TypeError,
-        UnicodeDecodeError,
-        ValueError,
-        json.JSONDecodeError,
-    ) as exc:
-        raise InvalidWorkspaceOverviewCursorError("Invalid workspace overview cursor") from exc
-    if not isinstance(workspace_id, str) or workspace_id == "":
-        raise InvalidWorkspaceOverviewCursorError("Invalid workspace overview cursor")
-    return _WorkspaceOverviewCursor(created_at=created_at, workspace_id=workspace_id)
 
 
 def _payloads_match(existing: Workspace, payload: WorkspaceCreateRequest) -> bool:
