+++ src/awf/mcp/server.py
@@ -12,21 +12,17 @@ cleanly when they show up alongside other MCP servers.
 
 from __future__ import annotations
 
+import inspect
 import json
-from types import SimpleNamespace
+from collections.abc import Awaitable, Callable
 from typing import Annotated, Any
 
-from fastapi import HTTPException
 from mcp.server.fastmcp import FastMCP
 from mcp.types import CallToolResult, TextContent
 from pydantic import Field
-from starlette.requests import Request
+from sqlalchemy.ext.asyncio import AsyncSession
 
-from awf.api.routes import artifacts as artifact_routes
-from awf.api.routes import merge_queue as merge_queue_routes
 from awf.api.routes import metrics as metrics_routes
-from awf.api.routes import validation as validation_routes
-from awf.api.routes import workspaces as workspace_routes
 from awf.api.schemas import (
     ErrorResponse,
     OperationListResponse,
@@ -39,7 +35,10 @@ from awf.api.schemas import (
 from awf.common.config import Settings, get_settings
 from awf.db.enums import AgentRuntime, OperationStatus, OperationType, TaskClass, WorkspaceStatus
 from awf.profiles.resolver import ProfileResolutionError
+from awf.service.artifacts import list_workspace_artifacts_metadata
 from awf.service.controls import WorkspaceControlError
+from awf.service.disk import DiskCheck
+from awf.service.merge_queue import InvalidMergeQueueCursorError, list_merge_queue_response
 from awf.service.metrics import (
     DEFAULT_FAILURE_EXAMPLE_LIMIT,
     DEFAULT_SUMMARY_WINDOW_HOURS,
@@ -47,11 +46,32 @@ from awf.service.metrics import (
     MAX_SUMMARY_WINDOW_HOURS,
     MIN_FAILURE_EXAMPLE_LIMIT,
     MIN_SUMMARY_WINDOW_HOURS,
+    summarize_failure_analysis_for_session,
+    summarize_resource_saturation_for_session,
+    summarize_slo_metrics_for_session,
+    summarize_workspace_reliability_for_session,
 )
+from awf.service.orphan_resources import OrphanResourceSummary
 from awf.service.overlap_graph import OverlapGraphQueueState, build_workspace_overlap_graph
+from awf.service.validation_provenance import list_validation_provenance_response
+from awf.service.workspace_observability import (
+    InvalidWorkspaceOverviewCursorError,
+    list_workspace_overview_response,
+    list_workspace_stale_reasons_response,
+)
+from awf.service.workspace_runtime_health import WorkspaceRuntimeHealthSummary
 from awf.service.workspaces import WorkspaceService
 
 StructuredToolResult = Annotated[CallToolResult, dict[str, Any]]
+DiskCheckProvider = Callable[[Settings], DiskCheck | Awaitable[DiskCheck]]
+OrphanResourceSummaryProvider = Callable[
+    [Settings, AsyncSession],
+    OrphanResourceSummary | Awaitable[OrphanResourceSummary],
+]
+RuntimeHealthSummaryProvider = Callable[
+    [Settings, AsyncSession, OrphanResourceSummary],
+    WorkspaceRuntimeHealthSummary | Awaitable[WorkspaceRuntimeHealthSummary],
+]
 
 # ── MCP tool registration ─────────────────────────────────────────────────
 
@@ -62,13 +82,9 @@ def build_mcp_server(
     name: str = "awf",
     instructions: str | None = None,
     settings: Settings | None = None,
-    disk_check_provider: metrics_routes.DiskCheckProvider | None = None,
-    orphan_resource_summary_provider: (
-        metrics_routes.OrphanResourceSummaryProvider | None
-    ) = None,
-    runtime_health_summary_provider: (
-        metrics_routes.RuntimeHealthSummaryProvider | None
-    ) = None,
+    disk_check_provider: DiskCheckProvider | None = None,
+    orphan_resource_summary_provider: OrphanResourceSummaryProvider | None = None,
+    runtime_health_summary_provider: RuntimeHealthSummaryProvider | None = None,
 ) -> FastMCP:
     """Construct a FastMCP instance with AWF's tools bound to ``service``.
 
@@ -395,16 +411,16 @@ def build_mcp_server(
         """Read-only operator observability: list the REST merge queue envelope."""
         async with service.session_factory() as session:
             try:
-                response = await merge_queue_routes.list_merge_queue(
+                response = await list_merge_queue_response(
+                    session,
                     repo_url=repo_url,
                     base_branch=base_branch,
                     workspace_status=workspace_status,
                     limit=limit,
                     cursor=cursor,
-                    session=session,
                 )
-            except HTTPException as exc:
-                return _http_exception_result(exc)
+            except InvalidMergeQueueCursorError:
+                return _error_result("INVALID_CURSOR", "Invalid merge queue cursor.")
         return _tool_result(response.model_dump(mode="json"))
 
     @mcp.tool(name="awf_list_workspace_overview")
@@ -418,63 +434,61 @@ def build_mcp_server(
         """Read-only operator observability: list the REST workspace overview envelope."""
         async with service.session_factory() as session:
             try:
-                response = await workspace_routes.list_workspace_overview(
+                response = await list_workspace_overview_response(
+                    session,
                     workspace_status=workspace_status,
                     agent=agent,
                     repo_url=repo_url,
                     limit=limit,
                     cursor=cursor,
-                    session=session,
                 )
-            except HTTPException as exc:
-                return _http_exception_result(exc)
+            except InvalidWorkspaceOverviewCursorError:
+                return _error_result("INVALID_CURSOR", "Invalid workspace overview cursor.")
         return _tool_result(response.model_dump(mode="json"))
 
     @mcp.tool(name="awf_list_workspace_validation")
     async def awf_list_workspace_validation(
         workspace_id: str = Field(..., description="Workspace ID to inspect."),
-    ) -> StructuredToolResult:
+    ) -> CallToolResult:
         """Read-only operator observability: list validation provenance for a workspace."""
         async with service.session_factory() as session:
-            try:
-                response = await validation_routes.list_validation_provenance(
-                    workspace_id,
-                    session=session,
-                )
-            except HTTPException as exc:
-                return _http_exception_result(exc)
+            response = await list_validation_provenance_response(
+                session,
+                workspace_id=workspace_id,
+            )
+            if response is None:
+                return _null_tool_result()
         return _tool_result(response.model_dump(mode="json"))
 
     @mcp.tool(name="awf_list_workspace_stale_reasons")
     async def awf_list_workspace_stale_reasons(
         workspace_id: str = Field(..., description="Workspace ID to inspect."),
         include_resolved: bool = Field(default=False),
-    ) -> StructuredToolResult:
+    ) -> CallToolResult:
         """Read-only operator observability: list structured workspace stale reasons."""
         async with service.session_factory() as session:
-            try:
-                response = await workspace_routes.list_workspace_stale_reasons(
-                    workspace_id,
-                    include_resolved=include_resolved,
-                    session=session,
-                )
-            except HTTPException as exc:
-                return _http_exception_result(exc)
+            response = await list_workspace_stale_reasons_response(
+                session,
+                workspace_id=workspace_id,
+                include_resolved=include_resolved,
+            )
+            if response is None:
+                return _null_tool_result()
         return _tool_result(response.model_dump(mode="json"))
 
     @mcp.tool(name="awf_list_workspace_artifacts")
     async def awf_list_workspace_artifacts(
         workspace_id: str = Field(..., description="Workspace ID to inspect."),
-    ) -> StructuredToolResult:
+    ) -> CallToolResult:
         """Read-only operator observability: list workspace artifact metadata only."""
         async with service.session_factory() as session:
-            try:
-                response = await artifact_routes.list_workspace_artifacts(
-                    workspace_id,
-                    session=session,
-                )
-            except HTTPException as exc:
-                return _http_exception_result(exc)
+            response = await list_workspace_artifacts_metadata(
+                session,
+                workspace_id=workspace_id,
+                work_dir=current_settings().work_dir,
+            )
+            if response is None:
+                return _null_tool_result()
         return _tool_result(response.model_dump(mode="json"))
 
     @mcp.tool(name="awf_get_failure_analysis_summary")
@@ -492,11 +506,12 @@ def build_mcp_server(
     ) -> StructuredToolResult:
         """Read-only operator observability: summarize workspace failure analysis."""
         async with service.session_factory() as session:
-            response = await metrics_routes.get_failure_analysis_summary(
+            summary = await summarize_failure_analysis_for_session(
+                session,
                 since_hours=since_hours,
-                limit=limit,
-                session=session,
+                failure_example_limit=limit,
             )
+        response = metrics_routes.FailureAnalysisSummaryResponse.model_validate(summary)
         return _tool_result(response.model_dump(mode="json"))
 
     @mcp.tool(name="awf_get_workspace_reliability_summary")
@@ -509,26 +524,42 @@ def build_mcp_server(
     ) -> StructuredToolResult:
         """Read-only operator observability: summarize workspace reliability metrics."""
         async with service.session_factory() as session:
-            response = await metrics_routes.get_workspace_reliability_summary(
-                since_hours=since_hours,
+            summary = await summarize_workspace_reliability_for_session(
+                session,
                 settings=current_settings(),
-                session=session,
+                since_hours=since_hours,
             )
+        response = metrics_routes.WorkspaceReliabilitySummaryResponse.model_validate(summary)
         return _tool_result(response.model_dump(mode="json"))
 
     @mcp.tool(name="awf_get_resource_saturation_summary")
     async def awf_get_resource_saturation_summary() -> StructuredToolResult:
         """Read-only operator observability: summarize resource saturation and admission."""
         async with service.session_factory() as session:
-            response = await metrics_routes.get_resource_saturation_summary(
-                request=_resource_saturation_request(
-                    disk_check_provider=disk_check_provider,
-                    orphan_resource_summary_provider=orphan_resource_summary_provider,
-                    runtime_health_summary_provider=runtime_health_summary_provider,
-                ),
-                settings=current_settings(),
+            settings_value = current_settings()
+            disk_check = await _provided_disk_check(
+                disk_check_provider=disk_check_provider,
+                settings=settings_value,
+            )
+            orphan_resources = await _provided_orphan_resources(
+                orphan_resource_summary_provider=orphan_resource_summary_provider,
+                settings=settings_value,
                 session=session,
             )
+            runtime_health = await _provided_runtime_health(
+                runtime_health_summary_provider=runtime_health_summary_provider,
+                settings=settings_value,
+                session=session,
+                orphan_resources=orphan_resources,
+            )
+            summary = await summarize_resource_saturation_for_session(
+                session,
+                settings=settings_value,
+                disk_check=disk_check,
+                orphan_resources=orphan_resources,
+                runtime_health=runtime_health,
+            )
+        response = metrics_routes.ResourceSaturationSummaryResponse.model_validate(summary)
         return _tool_result(response.model_dump(mode="json"))
 
     @mcp.tool(name="awf_get_slo_metrics_summary")
@@ -541,11 +572,12 @@ def build_mcp_server(
     ) -> StructuredToolResult:
         """Read-only operator observability: summarize SLO metrics."""
         async with service.session_factory() as session:
-            response = await metrics_routes.get_slo_metrics_summary(
-                since_hours=since_hours,
+            summary = await summarize_slo_metrics_for_session(
+                session,
                 settings=current_settings(),
-                session=session,
+                since_hours=since_hours,
             )
+        response = metrics_routes.SloMetricsSummaryResponse.model_validate(summary)
         return _tool_result(response.model_dump(mode="json"))
 
     @mcp.tool(name="awf_list_operations")
@@ -574,7 +606,7 @@ def build_mcp_server(
     @mcp.tool(name="awf_get_operation")
     async def awf_get_operation(
         operation_id: str = Field(..., description="Operation ID to inspect."),
-    ) -> StructuredToolResult:
+    ) -> CallToolResult:
         """Read-only operator observability: fetch one operation by id."""
         result = await service.get_operation(operation_id)
         if result is None:
@@ -629,19 +661,8 @@ def _tool_error(exc: WorkspaceControlError) -> CallToolResult:
     return _tool_result(error.model_dump(mode="json"), is_error=True)
 
 
-def _http_exception_result(exc: HTTPException) -> CallToolResult:
-    detail: object = exc.detail
-    if isinstance(detail, dict):
-        error_code = str(detail.get("error_code") or "HTTP_ERROR")
-        if exc.status_code == 404 and error_code == "NOT_FOUND":
-            return _null_tool_result()
-        error = ErrorResponse(
-            error_code=error_code,
-            message=str(detail.get("message") or "HTTP request failed."),
-            detail=detail.get("detail") if isinstance(detail.get("detail"), dict) else None,
-        )
-    else:
-        error = ErrorResponse(error_code="HTTP_ERROR", message=str(detail))
+def _error_result(error_code: str, message: str) -> CallToolResult:
+    error = ErrorResponse(error_code=error_code, message=message)
     return _tool_result(error.model_dump(mode="json"), is_error=True)
 
 
@@ -649,29 +670,48 @@ def _null_tool_result() -> CallToolResult:
     return CallToolResult(content=[], structuredContent=None)
 
 
-def _resource_saturation_request(
+async def _provided_disk_check(
     *,
-    disk_check_provider: metrics_routes.DiskCheckProvider | None,
-    orphan_resource_summary_provider: metrics_routes.OrphanResourceSummaryProvider | None,
-    runtime_health_summary_provider: metrics_routes.RuntimeHealthSummaryProvider | None,
-) -> Request:
-    state = SimpleNamespace()
-    if disk_check_provider is not None:
-        state.workspace_admission_disk_check = disk_check_provider
-    if orphan_resource_summary_provider is not None:
-        state.orphan_resource_summary_provider = orphan_resource_summary_provider
-    if runtime_health_summary_provider is not None:
-        state.runtime_health_summary_provider = runtime_health_summary_provider
-    app = SimpleNamespace(state=state)
-    return Request(
-        {
-            "type": "http",
-            "method": "GET",
-            "path": "/v1/metrics/resources/saturation",
-            "headers": [],
-            "app": app,
-        }
-    )
+    disk_check_provider: DiskCheckProvider | None,
+    settings: Settings,
+) -> DiskCheck | None:
+    if disk_check_provider is None:
+        return None
+    result = disk_check_provider(settings)
+    if inspect.isawaitable(result):
+        return await result
+    return result
+
+
+async def _provided_orphan_resources(
+    *,
+    orphan_resource_summary_provider: OrphanResourceSummaryProvider | None,
+    settings: Settings,
+    session: AsyncSession,
+) -> OrphanResourceSummary | None:
+    if orphan_resource_summary_provider is None:
+        return None
+    result = orphan_resource_summary_provider(settings, session)
+    if inspect.isawaitable(result):
+        return await result
+    return result
+
+
+async def _provided_runtime_health(
+    *,
+    runtime_health_summary_provider: RuntimeHealthSummaryProvider | None,
+    settings: Settings,
+    session: AsyncSession,
+    orphan_resources: OrphanResourceSummary | None,
+) -> WorkspaceRuntimeHealthSummary | None:
+    if runtime_health_summary_provider is None:
+        return None
+    if orphan_resources is None:
+        return None
+    result = runtime_health_summary_provider(settings, session, orphan_resources)
+    if inspect.isawaitable(result):
+        return await result
+    return result
 
 
 def _tool_result(payload: dict[str, Any], *, is_error: bool = False) -> CallToolResult:
